+++ backend/app/views/chat_rooms/chat_messages/create.json.jbuilder
